Add AnyBuffer drain and restructure any_buffer module
